Fixed bug with interfaces
When accounts were being logged into, the squeal interface would pop up
first, causing the bot to attempt to click it. Soon after logging in,
RuneScape would send us a different interface, but the bot would be in a
loop trying to click the spin button. I fixed it by checking if the
interface was open before clicking again.